Refatora exportação DXF/DWG e controles de LM/elementos
Refatora a lógica de exportação DXF/DWG para uma nova classe ExportacaoDxf, removendo o método exportar() do Program.cs. Adiciona controles para inserção da lista de materiais (LM) e elementos de obra baseados nos campos TCNM_CRIAR_LM e TCNM_LISTAR_PARAF do modelo, exibindo mensagens no console conforme a configuração. Atualiza o projeto para incluir ExportacaoDxf.cs e melhora a robustez da conexão com o modelo Tekla.
