+++ ConsoleTNKDxf/Program.cs
@@ -1,12 +1,6 @@
 ﻿using System;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using Tekla.Structures;
-using TSM = Tekla.Structures.Model;
-using TSO = Tekla.Structures.Model.Operations;
 using TSD = Tekla.Structures.Drawing;
-using System.Collections.Generic;
+using TSM = Tekla.Structures.Model;
 
 namespace ConsoleTNKDxf
 {
@@ -20,6 +14,12 @@ namespace ConsoleTNKDxf
             TSM.Model modelTemp = new TSM.Model();
 
 
+            if(!modelTemp.GetConnectionStatus())
+            {
+                Console.WriteLine("Não foi possível conectar ao modelo.");
+                return;
+            }
+
             string nomeModel = modelTemp.GetInfo().ModelName;
 
 
@@ -31,19 +31,9 @@ namespace ConsoleTNKDxf
             {
                 Console.WriteLine("Modelo corrente não possui desenho selecionado. Selecione o modelo e os desenhos a exportar no Document manager");
                 return;
-                //return new RespostaModelo(false, null, );
             }
 
-            
-            //if (!resposta.Sucesso)
-            //{
-            //    Console.WriteLine(resposta.Mensagem);
-            //    return;
-            //}
-
-            //TSM.Model model = resposta.Model;
-
-            exportar();
+            ExportacaoDxf.Exportar();
 
             IAdapterDesenho adapterDesenho = new AdapterDesenho(modelTemp);
             Console.ForegroundColor = ConsoleColor.Blue;
@@ -63,8 +53,6 @@ namespace ConsoleTNKDxf
             Console.WriteLine(@"Processo concluído com arquivos na pasta .\PlotFiles\Enviar, Pressione qualquer tecla para sair.");
 
             Console.ReadKey();
-
-            //Environment.Exit(0);
 
         }
 
@@ -192,93 +180,6 @@ namespace ConsoleTNKDxf
         //    return null; // Nenhum modelo correspondente encontrado
         //}
 
-        public static void exportar()
-        {
-            TSM.Model model = new TSM.Model();
-            string modelPath = model.GetInfo().ModelPath;
-
-            string xsplot = "";
-            TeklaStructuresSettings.GetAdvancedOption("XS_DRAWING_PLOT_FILE_DIRECTORY", ref xsplot);
-
-            var destino = modelPath + xsplot.Replace(".", "");
-
-
-            if (Directory.Exists(destino))
-            {
-                try
-                {
-                    var arquivosExistentes = Directory.GetFiles(destino, "*.dxf");
-                    foreach (var arquivo in arquivosExistentes)
-                    {
-                        try
-                        {
-                            File.Delete(arquivo);
-                        }
-                        catch { }
-                    }
-                }
-                catch { }
-            }
-
-
-
-            TSM.Operations.Operation.DisplayPrompt("Exporting DWG Files.");
-
-            string TSBinaryDir = "";
-
-            TSM.Model CurrentModel = new TSM.Model();
-
-            TeklaStructuresSettings.GetAdvancedOption("XSBIN", ref TSBinaryDir);
-
-
-            string ApplicationName = "Dwg.exe";
-
-            string ApplicationPath = Path.Combine(TSBinaryDir, "Applications\\Tekla\\Drawings\\DwgExport\\" + ApplicationName);
-
-            string dwgxportParams = "export outputDirectory=\"" + destino + "\"";
-
-
-            Process NewProcess = new Process();
-
-
-            if (File.Exists(ApplicationPath))
-            {
-
-                NewProcess.StartInfo.FileName = ApplicationPath;
-
-
-                try
-
-                {
-
-                    NewProcess.StartInfo.Arguments = dwgxportParams;
-
-                    NewProcess.Start();
-
-                    NewProcess.WaitForExit();
-
-                }
-
-                catch
-
-                {
-
-                    TSO.Operation.DisplayPrompt(ApplicationName + " failed to start.");
-
-                }
-
-            }
-
-            else
-
-            {
-
-                TSO.Operation.DisplayPrompt(ApplicationName + " not found.");
-
-            }
-
-            TSM.Operations.Operation.DisplayPrompt("DWG Files Exported.");
-
-        }
+        
     }
 }
